openspy/requirements.txt
SQLAlchemy
requests
pyyaml
munch
pymysql
python-keystoneclient
pika